% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_betas}
\alias{estimate_betas}
\title{Estimate modal values}
\usage{
estimate_betas(betas, method = "shorth")
}
\arguments{
\item{betas}{Matrix with dimensions n x m (parameters x iterations).}

\item{method}{Default value is shorth. See more in @details .}
}
\value{
Vector of estimated mode values for each parameter (n)
}
\description{
Estimates mode for each parameter in matrix given vector of values.
}
\details{
Available methods are "mfv", "lientz", "naive",
"venter", "grenander", "hsm", "hrm", "parzen", "tsybakov", and "asselin".
For more, see \code{\link[modeest]{mlv}} documentation.
}
\keyword{mode,parameter,estimation}
