% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_flip_probability}
\alias{estimate_flip_probability}
\title{Calculates misclassification probability for samples}
\usage{
estimate_flip_probability(flip.indicators, standardize = TRUE)
}
\arguments{
\item{flip.indicators}{Matrix of flip indicators (flip.cases or flip.controls)}

\item{standardize}{Flag to normalize probability values (only when they are not normalized). Recommended to keep default.}
}
\value{
Misclassification probabilities for samples
}
\description{
Given model defined by Shafquat et al (...), the function estimates the probability
that a case is not a true case or a control is not a true control with the
matrix of flip indicators (flip.cases or flip.controls) from plvm_lmm/plvm/gibbs/gibbs_lmm
}
\keyword{keywords}
