% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_beta}
\alias{get_beta}
\title{Return mode for a vector of values}
\usage{
get_beta(beta, method = "shorth")
}
\arguments{
\item{beta}{Vector of values..}

\item{method}{Default value is shorth. See more in @details .}
}
\value{
Estimated mode value.
}
\description{
Given vector of values beta, the function returns the mode as
estimated by \code{\link[modeest]{mlv}}.
}
\details{
Available methods are "mfv", "lientz", "naive",
"venter", "grenander", "hsm", "hrm", "parzen", "tsybakov", and "asselin".
For more, see \code{\link[modeest]{mlv}} documentation.
}
\keyword{keywords}
