% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logsum}
\alias{logsum}
\title{Log sum of values x and y}
\usage{
logsum(a, b)
}
\arguments{
\item{a}{Log value of x.}

\item{b}{Log value of y.}
}
\value{
Returns log(x + y).
}
\description{
Given a = log(x) and b = log(y), returns log(x + y).
}
\keyword{keywords}
