% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perturb_y}
\alias{perturb_y}
\title{Simulates misclassification in binary phenotype}
\usage{
perturb_y(y, lambda = NULL, c_alpha = NULL, seed = 1000)
}
\arguments{
\item{y}{Vector of values of 0s and 1s.}

\item{lambda}{Fraction of controls switched to cases. Default is NULL (no switch).}

\item{c_alpha}{Fraction of cases switched to controls. Default is NULL (no switch).}

\item{seed}{Seed number to ensure different switches take place everytime. Default is 1000.}
}
\value{
Vector with switched 0s and 1s according to defined lambda and alpha.
}
\description{
This function simulates misclassification of binary phenotype (vector y) by switching fraction lambda of 0 (controls)
to 1 (cases) and fraction c_alpha of 1 (cases) to 0 (controls).
}
\keyword{misclassification,phenotype,simulation}
