% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phelex.R
\name{phelex}
\alias{phelex}
\title{Estimate misclassification in phenotype}
\usage{
phelex(x, y, A = NULL, iterations = 1e+05, alpha.prior = c(1, 1),
  lambda.prior = c(1, 1), link = "pnorm", beta.prior = "norm",
  beta.prior.params = c(2, 1), beta.initial.vec = NULL,
  sigmaA.initial = 0.01, beta.iterations = 2e+05, beta.warmup = 20000,
  verbose = TRUE, stamp = 10000)
}
\arguments{
\item{x}{Genotype matrix with dimensions n x m.}

\item{y}{Phenotype vector with length n.}

\item{A}{Genetic similarity matrix with dimensions n x n.}

\item{iterations}{Total number of iterations}

\item{alpha.prior}{Beta prior parameters for true-positve rate}

\item{lambda.prior}{Beta prior parameters for false-positive rate}

\item{link}{probit or logistic model (options pnorm and plogis respectively)}

\item{beta.prior}{'norm'(default): Normal prior or 'unif' Uniform prior on fixed effects}

\item{beta.prior.params}{if beta.prior is norm, then (mean, sd), if unif then (min, max)}

\item{beta.initial.vec}{Initial values for fixed effects}

\item{sigmaA.initial}{Starting value for variance parameter sigmaA where u ~ MVN(0, sigmaA * A).}

\item{beta.iterations}{Number of iterations to sample fixed effects (should be greater than beta.warmup)}

\item{beta.warmup}{Burn-in/warmup for initial sampling for fixed effects}

\item{verbose}{Default TRUE. Prints progress information}

\item{stamp}{Iteration breakpoint to print time}
}
\value{
List containing
\itemize{
 \item betas: Matrix of estimated effect sizes for each SNP (SNPs[rows] x iterations[columns]).
 \item parameters: Matrix with estimated parameter values[rows] across iterations[columns].
 Order is c(sigmaA, pi1, pi2) where pi1/pi2 = false positive/false negative rates, sigmaA = variance parameter
 \item misclassified.cases: Matrix of alpha vectors where 1s represent false positives and 0s represent true positives as inferred at each iterations
 \item misclassified.controls: Matrix of lambda vectors where 1s represent false negatives and 0s represent true negatives as inferred at each iterations
 \item posterior: Vector of posterior probability across iterations
 \item accept: Vector of 1/0 values across iterations; 1 indicates proposal was accepted at iteration;0 o/w
 }
}
\description{
Estimates misclassification probabilities in observed GWAS phenotype y given genotypes dataset x.
The method follows the PheLEx algorithm to predict misclassification probabilities
using Adaptive Metropolis-Hastings and mixed model defined by Shafquat et al.
}
\keyword{hastings,mixed}
\keyword{metropolis}
\keyword{model}
\keyword{phelex,misclassification,GWAS,phenotype,adaptive}
