% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phelex_mh.R
\name{phelex_mh}
\alias{phelex_mh}
\title{Estimate misclassification in phenotype}
\usage{
phelex_mh(y, x, A, pi1.prior = c(1, 4), pi2.prior = c(1, 4),
  sigmaA.initial = 0.01, u.initial = NULL, beta.initial.vec = NULL,
  iterations = 10000, stamp = 1000, verbose = T)
}
\arguments{
\item{y}{Phenotype vector with length n.}

\item{x}{Genotype matrix with dimensions n x m.}

\item{A}{Genetic relatedness matrix with dimensions n x n.}

\item{pi1.prior}{hyperparameters for false positive rate pi1.}

\item{pi2.prior}{hyperparameters for false negative rate pi2.}

\item{sigmaA.initial}{Starting value for variance parameter sigmaA where u ~ MVN(0, sigmaA*A).}

\item{u.initial}{Starting values for random effects vector u. Default assumes random values.}

\item{beta.initial.vec}{Initial values for beta parameters in order c(Beta_a_1,...Beta_a_i). Default values are random values for all parameters.}

\item{iterations}{Number of iterations for sampling}

\item{stamp}{Iteration breakpoint to print time}

\item{verbose}{Default TRUE. Prints progress information}
}
\value{
List containing
\itemize{
 \item betas: Matrix of estimated effect sizes for each SNP (SNPs[rows] x iterations[columns]).
 \item parameters: Matrix with estimated parameter values[rows] across iterations[columns].
 Order is c(sigmaA, pi1, pi2) where pi1/pi2 = false positive/false negative rates, sigmaA = variance parameter
 \item misclassified.cases: Matrix of misclassification indicators where 1s represent false positives and 0s represent true positives as inferred at each iterations
 \item misclassified.controls: Matrix of misclassification indicators where 1s represent false negatives and 0s represent true negatives as inferred at each iterations
 }
}
\description{
Estimates misclassification probabilities in observed GWAS phenotype y given genotypes dataset x
The method follows the PheLEx-mh algorithm to predict misclassification probabilities
using Gibbs sampling and mixed model defined by Shafquat et al.
}
\keyword{model}
\keyword{phelex_mh,misclassification,GWAS,phenotype,gibbs,mixed}
