% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rekaya.R
\name{rekaya}
\alias{rekaya}
\title{Estimate misclassification in phenotype using Gibbs Sampler}
\usage{
rekaya(y, x, pi1.prior = c(1, 4), pi2.prior = c(1, 4),
  beta.initial.vec = NULL, iterations = 10000, stamp = 1000,
  verbose = T)
}
\arguments{
\item{y}{Phenotype vector with length n.}

\item{x}{Genotype matrix with dimensions n x i.}

\item{pi1.prior}{hyperparameters for false positive rate pi1. Default Beta(1, 4).}

\item{pi2.prior}{hyperparameters for false negative rate pi2. Default Beta(1, 4).}

\item{beta.initial.vec}{Initial values for beta parameters in order c(Beta_a_1,...Beta_a_i). Default values are random values for all parameters.}

\item{iterations}{Number of iterations for sampling}

\item{stamp}{Iteration breakpoint to print time}

\item{verbose}{Default TRUE. Prints progress information}
}
\value{
List containing
\itemize{
 \item betas: Matrix of estimated effect sizes for each SNP (SNPs[rows] x iterations[columns]).
 \item parameters: Matrix with estimated parameter values[rows] across iterations[columns].
 Order is c(mu, pi1, pi2) where mu is mean effect size value, pi1/pi2 false positive/false negative rates.
 \item misclassified.cases: Matrix of misclassification indicators where 1s represent false positives and 0s represent true positives as inferred at each iterations
 \item misclassified.controls: Matrix of misclassification indicators where 1s represent false negatives and 0s represent true negatives as inferred at each iterations
 }
}
\description{
Estimates misclassification probabilities in observed GWAS phenotype y given genotypes dataset x.
The method follows the algorithm defined by Rekaya et. al (PMC5138056; 2016) to predict misclassification probabilities
using Gibbs Sampling algorithm.
}
\keyword{rekaya,gibbs,misclassification,GWAS,phenotype}
