% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_na}
\alias{remove_na}
\title{Function to remove missing values from matrix}
\usage{
remove_na(x)
}
\arguments{
\item{x}{Genotype matrix with dimensions n x m}
}
\value{
Genotype matrix with dimension n x m with NA values replaced with
the corresponding column mean.
}
\description{
Replaces NA values from columns of matrix with column mean.
}
\keyword{replace,missing,impute}
