% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_phenotype}
\alias{get_phenotype}
\title{Computes modified phenotype}
\usage{
get_phenotype(misclassified.p.cases = NULL, misclassified.p.controls = NULL,
  y, case.threshold = NULL, control.threshold = NULL)
}
\arguments{
\item{misclassified.p.cases}{vector of misclassification probabilities for cases. If absent, only controls are switched}

\item{misclassified.p.controls}{vector of misclassification probabilities for controls. If absent, only cases are switched}

\item{y}{Vector of observed phenotype}

\item{case.threshold}{Threshold = percentile cut-off on misclassification probabilities. All cases with Pr(misclassification) > percentile threshold are switched to controls. (Default value: .99 i.e 99th percentile)}

\item{control.threshold}{Threshold = percentile cut-off on misclassification probabilities. All controls with Pr(misclassification) > percentile threshold are switched to cases. (Default value: .99 i.e 99th percentile)}
}
\value{
Modified Phenotype according to misclassification probabilities provided
}
\description{
Function computes corrected phenotype using misclassification probability for each sample.
}
\keyword{corrected,phenotype,misclassification}
