% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perturb_y}
\alias{perturb_y}
\title{Switches binary status of values}
\usage{
perturb_y(y, lambda = NULL, c_alpha = NULL, seed = 1000)
}
\arguments{
\item{y}{Vector of values of 0s and 1s.}

\item{lambda}{Fraction of controls switched to cases. Default is NULL (no switch).}

\item{c_alpha}{Fraction of cases switched to controls. Default is NULL (no switch).}

\item{seed}{Seed number to ensure different switches take place everytime. Default is 1000.}
}
\value{
Vector with switched 0s and 1s according to defined lambda and alpha.
}
\description{
For vector y, fraction lambda of 0s are changed to 1s and fraction alpha of 1s to 0s.
}
\keyword{keywords}
