% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phelex.R
\name{phelex}
\alias{phelex}
\title{Estimate misclassification in phenotype}
\usage{
phelex(x, y, A = NULL, beta.initial.vec = NULL, sigmaA.initial = 0.01,
  alpha.prior = c(1, 1), lambda.prior = c(1, 1), iterations = 1e+05,
  beta.prior.params = c(2, 1), beta.iterations = 2e+05, verbose = TRUE,
  stamp = 10000, link = "pnorm", beta.prior = "norm")
}
\arguments{
\item{x}{Genotype matrix with dimensions s x n.}

\item{y}{Phenotype vector with length s.}

\item{A}{Genetic similarity matrix with dimensions s x s.}

\item{beta.initial.vec}{Vector of initial values for beta parameters}

\item{sigmaA.initial}{Starting value for variance parameter sigmaA where u ~ MVN(0, sigmaA*A).}

\item{alpha.prior}{Vector for Beta prior on alpha i.e. true positive rate.}

\item{lambda.prior}{Vector for Beta prior on lambda i.e. false positive rate.}

\item{iterations}{Total number of iterations for Metropolis-Hastings Sampling}

\item{beta.prior.params}{if beta.prior is norm, then (mean, sd), if unif then (min, max)}

\item{beta.iterations}{Total number of iterations to sample beta values}

\item{verbose}{Default TRUE. Prints progress information}

\item{stamp}{Iteration breakpoint to print time}

\item{link}{probit or logistic model (options pnorm and plogis respectively)}

\item{beta.prior}{'norm'(default): Normal prior for Betas, 'unif' then uniform prior on Betas.}
}
\value{
List containing
\itemize{
 \item betas: Matrix of estimated effect sizes for each SNP (SNPs[rows] x iterations[columns]).
 \item parameters: Matrix with estimated parameter values[rows] across iterations[columns].
 Order is c(sigmaA, pi1, pi2) where pi1/pi2 = false positive/false negative rates, sigmaA = variance parameter
 \item misclassified.cases: Matrix of alpha vectors where 1s represent false positives and 0s represent true positives as inferred at each iterations
 \item misclassified.controls: Matrix of lambda vectors where 1s represent false negatives and 0s represent true negatives as inferred at each iterations
 \item posterior: Vector of posterior probability across iterations
 \item accept: Vector of 1/0 values across iterations; 1 indicates proposal was accepted at iteration;0 o/w
 }
}
\description{
Estimates misclassification probabilities in observed GWAS phenotype y given genotypes dataset x.
The method follows the PheLEx algorithm to predict misclassification probabilities
using Adaptive Metropolis-Hastings and mixed model defined by Shafquat et al.
}
\keyword{keywords}
