% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phelex_m.R
\name{phelex_m}
\alias{phelex_m}
\title{Estimate misclassification in phenotype}
\usage{
phelex_m(x, y, beta.initial.vec = NULL, alpha.prior = c(1, 1),
  lambda.prior = c(1, 1), iterations = 1e+05, beta.prior.params = c(1, 3),
  verbose = TRUE, stamp = 1000, mu.update = 0.5, link = "pnorm",
  beta.prior = "norm")
}
\arguments{
\item{x}{Genotype matrix with dimensions s x n.}

\item{y}{Phenotype vector with length s.}

\item{beta.initial.vec}{Vector of initial values for beta parameters i.e. effect sizes for n snps.}

\item{alpha.prior}{Vector for Beta prior on alpha.}

\item{lambda.prior}{Vector for Beta prior on lambda.}

\item{iterations}{Total number of iterations for Metropolis-Hastings Sampling}

\item{beta.prior.params}{if beta.prior is norm, then (mean, sd), if unif then (min, max)}

\item{verbose}{Default TRUE. Prints progress information}

\item{stamp}{Iteration breakpoint to print time}

\item{mu.update}{Fractions of iterations to start updating mean beta from.}

\item{link}{probit or logistic model (options pnorm and plogis respectively)}

\item{beta.prior}{'norm'(default): Normal prior for Betas, 'unif' then uniform prior on Betas.}
}
\value{
List containing
\itemize{
 \item betas: Matrix of estimated effect sizes for each SNP (SNPs[rows] x iterations[columns]).
 \item parameters: Matrix with estimated parameter values[rows] across iterations[columns].
 Order is c(mu, pi1, pi2) where pi1/pi2 = false positive/false negative rates, mu = mean effect size value
 \item misclassified.cases: Matrix of alpha vectors where 1s represent false positives and 0s represent true positives as inferred at each iterations
 \item misclassified.controls: Matrix of lambda vectors where 1s represent false negatives and 0s represent true negatives as inferred at each iterations
 \item posterior: Vector of posterior probability across iterations
 \item accept: Vector of 1/0 values across iterations; 1 indicates proposal was accepted at iteration;0 o/w
 }
}
\description{
Estimates misclassification probabilities in observed GWAS phenotype y given genotypes dataset x.
The method follows the PheLEx-m algorithm to predict misclassification probabilities
using Adaptive Metropolis-Hastings defined by Shafquat et al.
}
\keyword{keywords}
