% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_na}
\alias{remove_na}
\title{Function to remove missing values from matrix}
\usage{
remove_na(x)
}
\arguments{
\item{x}{Genotype matrix with dimensions s x n.}
}
\value{
Genotype matrix with dimension s x n with NA values replaced with
the corresponding column mean.
}
\description{
Replaces NA values from columns of matrix with column mean.
}
\keyword{keywords}
